move icon chip component location
