+++ packages/airview-compliance-ui/src/features/icon-chip/index.js
@@ -1,0 +1,1 @@
+export { IconChip } from "./icon-chip";
